requirements.txt
tiktoken>=0.5.1
openpyxl>=3.1.0
llama-cpp-python>=0.2.0
pandas>=2.0.0
requests>=2.28.0
